Check in 1.4.51

+++ src/extensions/localworker.cpp
@@ -566,9 +566,9 @@ void LocalWorker::configRlimit(RLimits *pRLimits, const XmlNode *pNode)
         ConfigCtx::getCurConfigCtx()->getLongValue(pNode, "CPUHardLimit", 0,
                 INT_MAX, 0));
     long memSoft = ConfigCtx::getCurConfigCtx()->getLongValue(pNode,
-                   "memSoftLimit", 0, INT_MAX, 0);
+                   "memSoftLimit", 0, LONG_MAX, 0);
     long memHard = ConfigCtx::getCurConfigCtx()->getLongValue(pNode,
-                   "memHardLimit", 0, INT_MAX, 0);
+                   "memHardLimit", 0, LONG_MAX, 0);
 
     if ((memSoft & (memSoft < 1024 * 1024)) ||
         (memHard & (memHard < 1024 * 1024)))
